requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.11.0
python-dotenv>=1.0.0
schedule>=1.2.0
